src/输入输出篇/接口服务/RPC服务/code/GRpc接口服务/C1/cli_sync.py
#!/usr/bin/env python
import grpc
from data_pb2_grpc import SquareServiceStub
from data_pb2 import Message
url = "localhost:5000"
conn = grpc.insecure_channel(url)
client = SquareServiceStub(channel=conn)
for i in client.rangeSquare(Message(message=12)):
    print(i.message)